🛠 refactor: 아바타 url CDN 주소 삭제
api gateway 적용에 따른 CDN 주소 삭제
