+++ src/auth/dtos/index.ts
@@ -1,3 +1,4 @@
+export * from './signup.req.dto';
+export * from './signup.res.dto';
 export * from './login.req.dto';
 export * from './login.res.dto';
-export * from './signup.req.dto';
