+++ src/auth/dtos/signup.res.dto.ts
@@ -1,0 +1,7 @@
+import { PickType } from '@nestjs/swagger';
+import { LoginResDto } from './login.res.dto';
+
+export class SignupResDto extends PickType(LoginResDto, [
+  'userId',
+  'profile',
+] as const) {}
