+++ src/auth/services/login.service.spec.ts
@@ -1,13 +1,13 @@
 import {
   GetUserByEmailOutboundPort,
   GetUserByEmailOutboundPortOutputDto,
-} from './../outbound-port/get-user-by-email.outbound-port';
+} from '../outbound-port/get-user-by-email.outbound-port';
 import { LoginService } from './login.service';
-import { User } from './../../user/entities/user.entity';
 import { UserFindByEmailService } from '../../user/services/user-find-by-email.service';
 import { Test } from '@nestjs/testing';
 import { BadRequestException } from '@nestjs/common';
 import * as bcrypt from 'bcrypt';
+import { User } from '../../entities/user.entity';
 
 class MockGetUserByEmailOutboundPort implements GetUserByEmailOutboundPort {
   private readonly users: GetUserByEmailOutboundPortOutputDto[];
