+++ src/auth/services/signup.service.spec.ts
@@ -1,14 +1,12 @@
-import { UserFindByEmailService } from './../../user/services/user-find-by-email.service';
 import { SignupService } from './signup.service';
-import { UserCreateService } from '../../user/services/user-create.service';
 import { ConflictException } from '@nestjs/common';
-import { User } from '../../user/entities/user.entity';
-import {
-  CreateUserOutboundPortInputDto,
-  CreateUserOutboundPort,
-} from '../../../dist/auth/outbound-port/create-user.outbound-port';
+import { User } from '../../entities/user.entity';
 import * as bcrypt from 'bcrypt';
-import { CheckEmailDuplicateOutboundPort } from '../../../dist/auth/outbound-port/check-email-duplicate.outbound-port';
+import {
+  CreateUserOutboundPort,
+  CreateUserOutboundPortInputDto,
+} from '../outbound-port/create-user.outbound-port';
+import { CheckEmailDuplicateOutboundPort } from '../outbound-port/check-email-duplicate.outbound-port';
 
 class MockCheckEmailDuplicateOutboundPort
   implements CheckEmailDuplicateOutboundPort
